requirements.txt
asgiref==3.2.7
dj-database-url==0.5.0
Django==3.0.5
djangorestframework==3.11.0
gunicorn==20.0.4
django-redis==4.11.0
psycopg2-binary==2.8.5
pytz==2019.3
redis==3.4.1
sqlparse==0.3.1
